dtv-server/pyproject.toml
[tool.poetry]
name = "dtv-server"
version = "0.1.0"
description = ""
authors = ["John McNelly <jkailimcnelly@gmail.com>"]
readme = "README.md"
packages = [{include = "dtv_server"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
pywin32 = "^304"
pillow = "^9.2.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
